+++ src/wave.cpp
@@ -22,6 +22,7 @@
 //' @param shift an optional logical value, if you would use a shift perturbation. See Details. Default is \code{FALSE}.
 //' @param toreBound a numeric value that specify the size of the grid. Default is -1.
 //' @param comment an optional logical value, indicating some informations during the execution. Default is \code{FALSE}.
+//' @param fixedSize an optional logical value, if you would impose a fixed sample size. Default is \code{TRUE}
 //'
 //' @details
 //' 
@@ -47,7 +48,9 @@
 //' where \eqn{p} is equal to the number of column of the matrix \code{X}.
 //' 
 //' For more informations on the option \code{shift}, see \code{\link{wpik}}.
-//'
+//' 
+//' If \code{fixedSize} is equal \code{TRUE}, the weakest associated vector is centered at each step of the algorithm. This ensures that the size of the selected sample is equal to the sum of the inclusion probabilities.
+//' 
 //' @return A vector of size \eqn{N} with elements equal 0 or 1. The value 1 indicates that the unit is selected while the value 0 is for non-chosen unit.
 //' 
 //' @author Raphaël Jauslin \email{raphael.jauslin@@unine.ch}
@@ -102,7 +105,8 @@ arma::vec wave(const arma::mat& X,
                bool tore = false,
                bool shift = false,
                double toreBound = -1,
-               bool comment = false) {
+               bool comment = false,
+               bool fixedSize = true) {
   // INITIALIZE CONSTANT
   double la1 = 1e+200;
   double la2 = 1e+200;
@@ -161,7 +165,7 @@ arma::vec wave(const arma::mat& X,
     
     arma::mat A(i_size,i_size);
     one = arma::ones<arma::mat>(i_size,1);
-    A = wpik(X.submat(i,ncol),re.elem(i),1.0,tore,shift,tb);
+    A = wpik(X.submat(i,ncol),re.elem(i),bound,tore,shift,tb);
     arma::mat D = diagmat(1/re.elem(i));
     A = A*D;
     
@@ -187,7 +191,9 @@ arma::vec wave(const arma::mat& X,
     }else{
         u = V.col(V.n_cols - 1);
     }
-    u = u - projOp(u,one);
+    if(fixedSize == true){
+      u = u - projOp(u,one);  
+    }
     
     la1 = 1e+200;
     la2 = 1e+200;
@@ -223,11 +229,23 @@ arma::vec wave(const arma::mat& X,
     if (i_size % 10 == 0){
       Rcpp::checkUserInterrupt();
     }
-      
+    
+    if(arma::sum(re.elem(i)) < (1-eps) && fixedSize == false){
+      Rcpp::Rcout << "The algorithm end because the remaining sum of inclusion probabilites is equal to : " << arma::sum(re.elem(i)) << 
+      "\nThe remaining inclusion probabilities are \n" << re.elem(i) <<
+      "\nBernoulli distribution are used to select the units.\n";
+      for(unsigned int tt = 0; tt < i.size(); tt++){
+        re[i[tt]] = R::rbinom(1,re[i[tt]]);
+      }
+      Rcpp::Rcout << re.elem(i) << "\n";
+      break;
+    }
+
     
   }
   
   if(comment  == true){
+    Rcpp::Rcout << re << std::endl;
     Rcpp::Rcout << "--- Sample selection finished ---" << std::endl;
   }
   
@@ -238,11 +256,28 @@ arma::vec wave(const arma::mat& X,
 
 /*** R
 
+
+N <- 36
+n <- 12
+x <- seq(1,sqrt(N),1)
+X <- as.matrix(cbind(runif(N),runif(N)))
+X <- as.matrix(expand.grid(x,x))
+pik <- rep(n/N,N)
+# pik <- sampling::inclusionprobabilities(runif(N),n)
+system.time(s <- wave(X,pik,tore = TRUE,shift = TRUE,fixedSize = TRUE))
+sum(s)
+plot(X)
+points(X[s == 1,],pch = 16)
+
+
+
 N <- 50
 n <- 15
 x <- as.matrix(runif(N),runif(N))
 pik <- sampling::inclusionprobabilities(runif(N),n)
-s <- wave(x,pik)
+s <- wave(x,pik,comment = TRUE,fixedSize = FALSE)
+sum(s)
+
 
 ############# EXAMPLE 1
 
@@ -257,7 +292,8 @@ plot(X)
 points(X[s == 1,],pch = 16)
 
 X <- as.matrix(cbind(runif(N),runif(N)))
-s <- wave(X,pik,tore = F,shift =F,comment = TRUE)
+s <- wave(X,pik,tore = F,shift =F,comment = TRUE,fixedSize = FALSE)
+sum(s)
 plot(X)
 points(X[s == 1,],pch = 16)
 
@@ -269,7 +305,7 @@ X <- as.matrix(cbind(rep(x,times = sqrt(N)),rep(x,each = sqrt(N))))
 pik <- rep(n/N,N)
 
 # W <- wpik(X,pik,bound = 1,tore = TRUE,shift = TRUE,toreBound = 15 )
-s <- wave(X,pik,tore = T,shift =T,comment = TRUE)
+s <- wave(X,pik,tore = T,shift =T,comment = TRUE,fixedSize = FALSE)
 plot(X)
 points(X[s == 1,],pch = 16)