+++ NEWS.md
@@ -1,5 +1,12 @@
 # NEWS for Package "WaveSampling"
 
+## WaveSampling 0.1.1
+
+### patch  
+
+  * add <arXiv:1910.13152> in DESCRIPTION file.
+  * add `fixedSize` option to choose if we would like to impose fixed sample size. Default is equal `TRUE`.
+  
 ## WaveSampling 0.1.0
 
 ### CRAN resubmission